:sparkles: Implement bloc logic for updating and submitting body weight
- Implement bloc logic for updating body weight in the home page
- Implement bloc logic for submitting body weight in the home page

This commit adds the functionality to update and submit body weight in the home page using the bloc pattern. The user can now enter their body weight in kilograms and submit it, triggering the appropriate bloc events and state changes. This enhances the user experience and interaction with the PortionControl app.

Signed-off-by: Dmytro Turskyi <dmytro.turskyi@gmail.com>